Disable reserved pin and peripheral on odin module.
Suppose there is a module that is using a PWM channel for an onboard peripheral.  It's critical that the associated timer peripheral configuration not be modified by the user application.  It's also critical that the associated GPIO pin not be modified by the user application.

This takes two steps:  We can prevent the user application from using the associated timer by removing it from PeripheralNames.h and PeripheralPins.c.  This will mean that the peripheral instance structure cannot be found via pinmap.  Any developer using a reserved peripheral should already know which pin/peripheral combination to use and this need not be reusable code, since it is very clearly target-specific.

Currently the removal of specific peripherals is accomplished using preprocessor macros, however in the long term, this should be accomplished using code generation.  The exact format of peripheral configuration in yotta config is not finalized, so only the ```reserved``` config key is present, since it is sufficient to disable a peripheral using the preprocessor.  A more generic solution is needed.

Next, we need to remove the pin’s name from the PinName enum.  Currently, pin names defined for a chip are defined in the chip-level HAL.  These should be defined in the chip-level target, so that yotta-config can override them.  When that change is made, adding a pin name to the reserved pins list will cause it to be removed from the pins list.  The pin name code generator will then produce two enums: ```PinName``` and ```ReservedPinName```.

Any pin that is present in the pins key of ```/hardware/reserved``` will be removed from the pins section before generating the ```PinName``` enum.  For example:

Chip level target config data:

```JSON
"hardware" : {
  "pins" : {
    "PZ_31" : "0xA5A5",
    "PZ_32" : "0x1234"
  }
}
```

Board/module level target config data:

```JSON
"hardware" : {
  "reserved" : {
    "pins" {
       "PZ_32" : true
    }
  }
}
```

When the two config blocks are merged by yotta, they will generate:

```JSON
"hardware" : {
  "pins" : {
    "PZ_31" : "0xA5A5",
    "PZ_32" : "0x1234"
  },
  "reserved" : {
    "pins" {
       "PZ_32" : true
    }
  }
}
```

This will cause the pin name generator script to output:

```C++
typedef enum {
    PZ_31 = 0xA5A5,
} PinName;
typedef enum {
    PZ_32 = 0x1234,
} ReservedPinName;
```

Note that a static cast allows a developer of a reserved pin/peripheral driver to still use existing functionality.  There is also the possibility that we could add a ```reservedPinMap()``` to permit reserved module developers to use mbed APIs.

This commit does not add all of these features.  It does, however, provide the configuration required to allow them to be added later.